refactor : AbstractLottery 추상클래스 추가 하여 공통적인 부분 추출
