+++ src/main/java/step2/lottery/Lottery.java
@@ -5,33 +5,14 @@ import step2.winning.Winning;
 
 import java.util.Set;
 
-public class Lottery {
-    private final Set<LotteryNumber> lottoNumbers;
-
+public class Lottery extends AbstractLottery {
     public Lottery(Set<LotteryNumber> lottoNumbers) {
-        this.lottoNumbers = lottoNumbers;
+        super(lottoNumbers);
     }
+
     public Winning matchWith(PrizeWinningNumber prize) {
         int matchCount = prize.matchCount(lottoNumbers);
         boolean bonus = (matchCount == 5 && prize.isMatchBounus(lottoNumbers));
         return Winning.valueOf(matchCount, bonus);
-    }
-
-    public Set<LotteryNumber> getLottoNumbers() {
-        return lottoNumbers;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj instanceof Lottery) {
-            Lottery other = (Lottery) obj;
-            return lottoNumbers.equals(other.lottoNumbers);
-        }
-        return false;
-    }
-
-    @Override
-    public int hashCode() {
-        return lottoNumbers.hashCode();
     }
 }
